- 새로운 유저 db를 이제 마이페이지에서 조회 - 그룹컴포넌트 코드 정리/분산
