+++ front/swit/src/components/myPage/MyProfileComponent.js
@@ -1,62 +1,79 @@
 import { useState, useEffect } from 'react';
 import { API_SERVER_HOST, getUserProfile, putUserProfile, postUserImage, getUserImage } from '../../api/UserApi';
+// import { info } from '../../api/loginApi';
 
 const initState = {
-    user_name: '',
-    user_nick: '',
-    user_phone: '',
-    user_email: '',
-    user_password: '',
-    user_image: ''
+    userName: '',
+    userNick: '',
+    userPhone: '',
+    userEmail: '',
+    userPassword: '',
+    userImage: ''
 }
 
 const host = API_SERVER_HOST
 
-const MyProfileComponent = ({ user_id }) => {
+const MyProfileComponent = ({ userId }) => {
     const [isModalOpen, setIsModalOpen] = useState(false)
-    const [user, setUser] = useState(initState)             //최종 반영된 마이페이지 프로필
-    const [modalUser, setModalUser] = useState(initState)   //모달창에서만 보이는 임시 프로필 정보
-    const [userImage, setUserImage] = useState(null)        //이미지 관리
-    //console.info("ccccccc")
+    const [user, setUser] = useState(initState)             // 최종 반영된 마이페이지 프로필
+    const [modalUser, setModalUser] = useState(initState)   // 모달창에서만 보이는 임시 프로필 정보
+    const [userImage, setUserImage] = useState(null)        // 이미지 관리
+    const [isSocialLogin, setIsSocialLogin] = useState(false); // 소셜 로그인 여부 확인
 
+    // 최초로 페이지 로드될 때 사용자 정보 가져오기
     useEffect(() => {
-        const fetchUserProfile = async () => {
+        const fetchUserInfo = async () => {
             try {
-                const data = await getUserProfile(user_id);
-                setUser(data);
-                setModalUser({ ...data });
-                const imageUrl = await getUserImage(user_id);
+                // const data = await info();
+                const userData = await getUserProfile(userId);
+                setUser(userData);
+                setModalUser({ ...userData });
+                const imageUrl = await getUserImage(userId);
                 setUserImage(imageUrl);
+
+                // 소셜 로그인 여부 확인
+                if (userData && (userData.userSnsConnect === 'NAVER' || userData.userSnsConnect === 'KAKAO')) {
+                    console.log(userId + '님은 소셜 로그인 유저입니다.');
+                    setIsSocialLogin(true);
+                }
             } catch (error) {
                 console.error(error);
             }
         };
-        fetchUserProfile();
-    }, [user_id])
-    //console.info("ddddddd")
+        fetchUserInfo();
+    }, [userId]);
 
     const openModal = () => {
+        if (isSocialLogin) {
+            console.log(userId + '님은 프로필 수정을 할 수 없습니다.');
+            alert('소셜 로그인 계정은 프로필 수정을 할 수 없습니다.');
+            return;
+        }
         setModalUser({ ...user }); // 모달창을 열 때 현재 프로필 데이터 복사
         setIsModalOpen(true);
-    }
+    };
 
-    //수정하기 버튼으로 모달창 닫을 때 이벤트
+    // 수정하기 버튼으로 모달창 닫을 때 이벤트
     const closeModal = () => {
         setIsModalOpen(false);
-    }
+    };
 
-    //x 버튼으로 모달창 닫을 때 이벤트
+    // x 버튼으로 모달창 닫을 때 이벤트
     const closeModalOptionX = () => {
         setIsModalOpen(false);
-        //setUserImage(null); // 모달이 닫힐 때 새 이미지 state 초기화
-    }
+        // setUserImage(null); // 모달이 닫힐 때 새 이미지 state 초기화
+    };
 
     const handleTextChange = (e) => {
         const { name, value } = e.target;
         setModalUser((prev) => ({ ...prev, [name]: value }));
-    }
+    };
 
     const handleImageUpload = async (e) => {
+        if (isSocialLogin) {
+            alert('소셜 로그인 계정은 프로필 수정을 할 수 없습니다.');
+            return;
+        }
         const img = e.target.files[0];
         if (img) {
             const reader = new FileReader();
@@ -67,32 +84,36 @@ const MyProfileComponent = ({ user_id }) => {
 
             // 이미지 업로드 처리
             const formData = new FormData();
-            formData.append("user_image", img);
+            formData.append("userImage", img);
             try {
-                await postUserImage(user_id, formData);
+                await postUserImage(userId, formData);
                 // 프로필 이미지 정보 갱신(url로 관리)
-                const imageUrl = await getUserImage(user_id);
+                const imageUrl = await getUserImage(userId);
                 setUserImage(imageUrl);
             } catch (error) {
                 console.error('Error uploading image:', error);
             }
         }
-    }
+    };
 
     const handleClickModify = async () => {
+        if (isSocialLogin) {
+            alert('소셜 로그인 계정은 프로필 수정을 할 수 없습니다.');
+            return;
+        }
         try {
             // 텍스트 수정 처리
             await putUserProfile(modalUser);
 
             // 프로필 텍스트 정보 갱신
-            const updatedProfile = await getUserProfile(user_id);
+            const updatedProfile = await getUserProfile(userId);
             setUser(updatedProfile);
 
             closeModal();
         } catch (error) {
             console.error('Error updating profile:', error);
         }
-    }
+    };
 
     return (
         <>
@@ -116,19 +137,19 @@ const MyProfileComponent = ({ user_id }) => {
                     <div className="flex-grow">
                         <div className="mb-2">
                             <span className="text-gray-500">이름: </span>
-                            <span className="font-bold text-black">{user.user_name}</span>
+                            <span className="font-bold text-black">{user.userName}</span>
                         </div>
                         <div className="mb-2">
                             <span className="text-gray-500">닉네임: </span>
-                            <span className="font-bold text-black">{user.user_nick}</span>
+                            <span className="font-bold text-black">{user.userNick}</span>
                         </div>
                         <div className="mb-2">
                             <span className="text-gray-500">전화번호: </span>
-                            <span className="font-bold text-black">{user.user_phone}</span>
+                            <span className="font-bold text-black">{user.userPhone}</span>
                         </div>
                         <div className="mb-2">
                             <span className="text-gray-500">이메일: </span>
-                            <span className="font-bold text-black">{user.user_email}</span>
+                            <span className="font-bold text-black">{user.userEmail}</span>
                         </div>
                     </div>
                     <button
@@ -164,22 +185,22 @@ const MyProfileComponent = ({ user_id }) => {
                         <div className="mb-4">
                             <label className="block mb-1">이름:</label>
                             <input className="w-full border border-gray-300 p-2 rounded"
-                                name="user_name" type="text" value={modalUser.user_name} onChange={handleTextChange} />
+                                name="userName" type="text" value={modalUser.userName} onChange={handleTextChange} />
                         </div>
                         <div className="mb-4">
                             <label className="block mb-1">닉네임:</label>
                             <input className="w-full border border-gray-300 p-2 rounded"
-                                name="user_nick" type="text" value={modalUser.user_nick} onChange={handleTextChange} />
+                                name="userNick" type="text" value={modalUser.userNick} onChange={handleTextChange} />
                         </div>
                         <div className="mb-4">
                             <label className="block mb-1">전화번호:</label>
                             <input className="w-full border border-gray-300 p-2 rounded"
-                                name="user_phone" type="text" value={modalUser.user_phone} onChange={handleTextChange} />
+                                name="userPhone" type="text" value={modalUser.userPhone} onChange={handleTextChange} />
                         </div>
                         <div className="mb-4">
                             <label className="block mb-1">이메일:</label>
                             <input className="w-full border border-gray-300 p-2 rounded"
-                                name="user_email" type="text" value={modalUser.user_email} onChange={handleTextChange} />
+                                name="userEmail" type="text" value={modalUser.userEmail} onChange={handleTextChange} />
                         </div>
                         <button className="bg-red-500 text-white px-4 py-2 rounded w-full" onClick={handleClickModify}>수정하기</button>
                     </div>
